# backend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "AeroFS API",
    "description": "Remote file manager REST API. All endpoints require session authentication except public share endpoints.",
    "license": {
      "name": "Proprietary"
    },
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/v1/connections": {
      "get": {
        "tags": [
          "connections"
        ],
        "summary": "List all available connections from database (scoped to user's permissions)",
        "operationId": "list_connections",
        "responses": {
          "200": {
            "description": "List of storage connections",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ConnectionDetailResponse"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "CookieAuth": []
          },
          {
            "BearerAuth": []
          }
        ]
      },
      "post": {
        "tags": [
          "connections"
        ],
        "summary": "Create a new connection with encrypted credential storage (Admin only, Fail-Closed Transactional)",
        "operationId": "create_connection",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateConnectionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Connection created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreateConnectionResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden - Admin required",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "Conflict - Connection name already exists",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "CookieAuth": []
          },
          {
            "BearerAuth": []
          }
        ]
      }
    },
...
    "/api/v1/connections/{connection_id}/search": {
      "get": {
        "tags": [
          "search"
        ],
        "summary": "Recursive search files in a connection",
        "operationId": "search_files",
        "parameters": [
          {
            "name": "connection_id",
            "in": "path",
            "description": "Connection ID",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "path",
            "in": "query",
            "required": false,
            "schema": {
              "type": [
                "string",
                "null"
              ]
            }
          },
          {
            "name": "query",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "regex",
            "in": "query",
            "required": false,
            "schema": {
              "type": [
                "boolean",
                "null"
              ]
            }
          },
          {
            "name": "max_depth",
            "in": "query",
            "required": false,
            "schema": {
              "type": [
                "integer",
                "null"
              ],
              "minimum": 0
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": [
                "integer",
                "null"
              ],
              "minimum": 0
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Search results",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SearchOutput"
                }
              }
            }
          },
          "400": {
            "description": "Bad request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "CookieAuth": []
          },
          {
            "BearerAuth": []
          }
        ]
      }
    },
    "/api/v1/connections/{id}": {
      "get": {
        "tags": [
          "connections"
        ],
        "summary": "Get a specific connection and its capabilities",
        "operationId": "get_connection",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Connection identifier",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Connection details and capabilities",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConnectionDetailResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Connection not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "CookieAuth": []
          },
          {
            "BearerAuth": []
          }
        ]
      },
      "put": {
        "tags": [
          "connections"
        ],
        "summary": "Update an existing connection (Admin only, with atomic hot-swap)",
        "operationId": "update_connection",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Connection identifier",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateConnectionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Connection updated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConnectionActionResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden - Admin required",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Connection not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "CookieAuth": []
          },
          {
            "BearerAuth": []
          }
        ]
      },
      "delete": {
        "tags": [
          "connections"
        ],
        "summary": "Delete a connection",
        "operationId": "delete_connection",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Connection identifier",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Connection deleted successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConnectionActionResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden - Admin required",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Connection not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "CookieAuth": []
          },
          {
            "BearerAuth": []
          }
        ]
      }
    },
...
    "/api/v1/connections/{id}/test": {
      "post": {
        "tags": [
          "connections"
        ],
        "summary": "Test connection connectivity",
        "operationId": "test_connection",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Connection identifier",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Connection test completed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TestConnectionResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Connection not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "CookieAuth": []
          },
          {
            "BearerAuth": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Capabilities": {
        "type": "object",
        "description": "Describes the operational capabilities supported by a specific VFS provider instance.",
        "required": [
          "list",
          "stat",
          "read",
          "write",
          "create_file",
          "create_dir",
          "delete",
          "rename",
          "copy",
          "move_",
          "upload",
          "download",
          "resume_upload",
          "resume_download",
          "range_read",
          "resumable_read",
          "multipart_write",
          "resumable_write",
          "write_can_append",
          "write_can_empty",
          "write_can_multi",
          "presign_read",
          "presign_write",
          "conditional_write",
          "atomic_write",
          "atomic_rename",
          "server_side_copy",
          "native_copy",
          "symlink",
          "permissions",
          "watch",
          "checksum",
          "native_checksum",
          "checksums",
          "computed_checksums"
        ],
        "properties": {
          "atomic_rename": {
            "type": "boolean"
          },
          "atomic_write": {
            "type": "boolean"
          },
          "checksum": {
            "type": "boolean"
          },
          "checksums": {
            "$ref": "#/components/schemas/ChecksumCapabilities"
          },
          "computed_checksums": {
            "$ref": "#/components/schemas/ChecksumCapabilities"
          },
          "conditional_write": {
            "type": "boolean"
          },
          "copy": {
            "type": "boolean"
          },
          "create_dir": {
            "type": "boolean"
          },
          "create_file": {
            "type": "boolean"
          },
          "delete": {
            "type": "boolean"
          },
          "download": {
            "type": "boolean"
          },
          "list": {
            "type": "boolean"
          },
          "move_": {
            "type": "boolean"
          },
          "multipart_write": {
            "type": "boolean"
          },
          "native_checksum": {
            "type": "boolean"
          },
          "native_copy": {
            "type": "boolean"
          },
          "permissions": {
            "type": "boolean"
          },
          "presign_read": {
            "type": "boolean"
          },
          "presign_write": {
            "type": "boolean"
          },
          "range_read": {
            "type": "boolean"
          },
          "read": {
            "type": "boolean"
          },
          "rename": {
            "type": "boolean"
          },
          "resumable_read": {
            "type": "boolean"
          },
          "resumable_write": {
            "type": "boolean"
          },
          "resume_download": {
            "type": "boolean"
          },
          "resume_upload": {
            "type": "boolean"
          },
          "server_side_copy": {
            "type": "boolean"
          },
          "stat": {
            "type": "boolean"
          },
          "symlink": {
            "type": "boolean"
          },
          "upload": {
            "type": "boolean"
          },
          "watch": {
            "type": "boolean"
          },
          "write": {
            "type": "boolean"
          },
          "write_can_append": {
            "type": "boolean"
          },
          "write_can_empty": {
            "type": "boolean"
          },
          "write_can_multi": {
            "type": "boolean"
          }
        }
      },
      "ChecksumCapabilities": {
        "type": "object",
        "description": "Typed granular checksum capabilities for storage backends",
        "required": [
          "md5",
          "crc32",
          "crc32c",
          "sha1",
          "sha256"
        ],
        "properties": {
          "crc32": {
            "type": "boolean"
          },
          "crc32c": {
            "type": "boolean"
          },
          "md5": {
            "type": "boolean"
          },
          "sha1": {
            "type": "boolean"
          },
          "sha256": {
            "type": "boolean"
          }
        }
      },
...
      "Connection": {
        "type": "object",
        "required": [
          "id",
          "name",
          "provider",
          "base_path",
          "read_only",
          "enabled",
          "status",
          "created_at",
          "updated_at"
        ],
        "properties": {
          "base_path": {
            "type": "string"
          },
          "created_at": {
            "type": "string",
            "format": "date-time"
          },
          "enabled": {
            "type": "boolean"
          },
          "error_message": {
            "type": [
              "string",
              "null"
            ]
          },
          "host": {
            "type": [
              "string",
              "null"
            ]
          },
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "port": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int32",
            "minimum": 0
          },
          "provider": {
            "$ref": "#/components/schemas/ProviderKind"
          },
          "read_only": {
            "type": "boolean"
          },
          "status": {
            "$ref": "#/components/schemas/ConnectionStatus"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time"
          },
          "username": {
            "type": [
              "string",
              "null"
            ]
          }
        }
      },
      "ConnectionActionResponse": {
        "type": "object",
        "required": [
          "success",
          "message"
        ],
        "properties": {
          "message": {
            "type": "string"
          },
          "success": {
            "type": "boolean"
          }
        }
      },
      "ConnectionDetailResponse": {
        "type": "object",
        "required": [
          "connection",
          "capabilities"
        ],
        "properties": {
          "capabilities": {
            "$ref": "#/components/schemas/Capabilities"
          },
          "connection": {
            "$ref": "#/components/schemas/Connection"
          }
        }
      },
...
      "ConnectionStatus": {
        "type": "string",
        "enum": [
          "disconnected",
          "connecting",
          "connected",
          "reconnecting",
          "failed"
        ]
      },
      "CreateConnectionRequest": {
        "type": "object",
        "required": [
          "name",
          "provider"
        ],
        "properties": {
          "base_path": {
            "type": [
              "string",
              "null"
            ]
          },
          "host": {
            "type": [
              "string",
              "null"
            ]
          },
          "name": {
            "type": "string"
          },
          "port": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int32",
            "minimum": 0
          },
          "provider": {
            "$ref": "#/components/schemas/ProviderKind"
          },
          "read_only": {
            "type": [
              "boolean",
              "null"
            ]
          },
          "secret": {
            "type": [
              "string",
              "null"
            ]
          },
          "username": {
            "type": [
              "string",
              "null"
            ]
          }
        }
      },
      "CreateConnectionResponse": {
        "type": "object",
        "required": [
          "success",
          "id",
          "message"
        ],
        "properties": {
          "id": {
            "type": "string"
          },
          "message": {
            "type": "string"
          },
          "success": {
            "type": "boolean"
          }
        }
      },
...
      "ErrorCategory": {
        "type": "string",
        "enum": [
          "validation",
          "authentication",
          "authorization",
          "not_found",
          "conflict",
          "transfer_cancelled",
          "payload_too_large",
          "insufficient_storage",
          "rate_limited",
          "timeout",
          "provider",
          "io",
          "security",
          "internal"
        ]
      },
...
      "ErrorDetail": {
        "type": "object",
        "required": [
          "code",
          "category",
          "retryable",
          "message"
        ],
        "properties": {
          "category": {
            "$ref": "#/components/schemas/ErrorCategory"
          },
          "code": {
            "type": "string"
          },
          "details": {},
          "message": {
            "type": "string"
          },
          "retryable": {
            "type": "boolean"
          },
          "user_action": {
            "type": [
              "string",
              "null"
            ]
          }
        }
      },
      "ErrorResponse": {
        "type": "object",
        "required": [
          "error"
        ],
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ErrorDetail"
          }
        }
      },
...
      "FileEntry": {
        "type": "object",
        "required": [
          "name",
          "path",
          "kind",
          "is_hidden"
        ],
        "properties": {
          "is_hidden": {
            "type": "boolean"
          },
          "kind": {
            "$ref": "#/components/schemas/FileKind"
          },
          "mime_type": {
            "type": [
              "string",
              "null"
            ]
          },
          "modified_at": {
            "type": [
              "string",
              "null"
            ],
            "format": "date-time"
          },
          "name": {
            "type": "string"
          },
          "path": {
            "type": "string"
          },
          "permissions": {
            "type": [
              "string",
              "null"
            ]
          },
          "size": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int64",
            "minimum": 0
          },
          "symlink_target": {
            "type": [
              "string",
              "null"
            ]
          }
        }
      },
      "FileKind": {
        "type": "string",
        "enum": [
          "file",
          "directory",
          "symlink"
        ]
      },
...
      "ProviderKind": {
        "type": "string",
        "enum": [
          "local",
          "ftp",
          "ftps",
          "sftp",
          "s3"
        ]
      },
...
      "SearchOutput": {
        "type": "object",
        "required": [
          "results",
          "truncated",
          "total_scanned",
          "errors"
        ],
        "properties": {
          "errors": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "results": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/FileEntry"
            }
          },
          "total_scanned": {
            "type": "integer",
            "minimum": 0
          },
          "truncated": {
            "type": "boolean"
          }
        }
      },
...
      "TestConnectionResponse": {
        "type": "object",
        "required": [
          "success",
          "latency_ms",
          "message"
        ],
        "properties": {
          "latency_ms": {
            "type": "integer",
            "format": "int64",
            "minimum": 0
          },
          "message": {
            "type": "string"
          },
          "success": {
            "type": "boolean"
          }
        }
      },
...
      "UpdateConnectionRequest": {
        "type": "object",
        "properties": {
          "base_path": {
            "type": [
              "string",
              "null"
            ]
          },
          "enabled": {
            "type": [
              "boolean",
              "null"
            ]
          },
          "host": {
            "type": [
              "string",
              "null"
            ]
          },
          "name": {
            "type": [
              "string",
              "null"
            ]
          },
          "port": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int32",
            "minimum": 0
          },
          "read_only": {
            "type": [
              "boolean",
              "null"
            ]
          },
          "secret": {
            "type": [
              "string",
              "null"
            ]
          },
          "username": {
            "type": [
              "string",
              "null"
            ]
          }
        }
      },
...
    },
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      },
      "CookieAuth": {
        "type": "apiKey",
        "in": "cookie",
        "name": "aerofs_session"
      }
    }
  },
...
}